requirements.txt
# Core PDF extraction
pdfplumber
pypdf

# OCR (optional — requires Tesseract and Poppler installed on the system)
pytesseract
pdf2image
Pillow

# Web UI
flask

# Azure Document Intelligence (optional — only needed when using --enable-azure)
azure-ai-formrecognizer>=3.3.0

# Testing
pytest
